门店商品补充平台分类条件

+++ jmp/src/main/java/com/vpu/mp/service/pojo/shop/store/goods/StoreGoodsListQueryParam.java
@@ -15,7 +15,7 @@ import lombok.Data;
 @Data
 public class StoreGoodsListQueryParam {
 	
-	private Integer catId;
+	private Short catId;
 	private Byte isOnSale;
 	
 	/**
